InitialWalletView Tests (#1199)
